+++ crates/core/src/strategy/reflink.rs
@@ -1,7 +1,43 @@
+use super::{Strategy, StrategyInit};
 use crate::{Error, InitProgress, Result};
 use std::fs;
 use std::path::Path;
 use walkdir::WalkDir;
+
+pub(super) struct LinuxReflinkStrategy;
+
+impl Strategy for LinuxReflinkStrategy {
+    fn copy_directory(&self, from: &Path, to: &Path) -> Result<()> {
+        let destination_parent = same_filesystem_parent(from, to)?;
+        verify_reflinks_linux(destination_parent)?;
+        clone_directory_linux(from, to)
+    }
+
+    fn initialize_directory(
+        &self,
+        path: &Path,
+        _progress: &mut dyn FnMut(InitProgress),
+    ) -> Result<StrategyInit> {
+        verify_reflinks_linux(path)?;
+        Ok(StrategyInit::AlreadyNative)
+    }
+}
+
+fn same_filesystem_parent<'a>(from: &Path, to: &'a Path) -> Result<&'a Path> {
+    use std::os::unix::fs::MetadataExt;
+
+    let destination_parent = to
+        .parent()
+        .ok_or_else(|| Error::Path(format!("destination has no parent: {}", to.display())))?;
+    if fs::metadata(from)?.dev() == fs::metadata(destination_parent)?.dev() {
+        Ok(destination_parent)
+    } else {
+        Err(Error::CowUnavailable(format!(
+            "Linux reflinks require source and destination on the same filesystem: {}",
+            to.display()
+        )))
+    }
+}
 
 pub(super) fn clone_directory_linux(from: &Path, to: &Path) -> Result<()> {
     fs::create_dir(to)?;
@@ -81,6 +117,25 @@ pub(super) fn reflink_file_linux(from: &Path, to: &Path) -> Result<()> {
         from.display(),
         std::io::Error::last_os_error()
     )))
+}
+
+pub(super) fn verify_reflinks_linux(path: &Path) -> Result<()> {
+    let operation_id = ulid::Ulid::new();
+    let source = path.join(format!(".rift-reflink-probe-{operation_id}"));
+    let destination = path.join(format!(".rift-reflink-probe-copy-{operation_id}"));
+    fs::write(&source, b"rift")?;
+    let result = reflink_file_linux(&source, &destination).map_err(|error| match error {
+        Error::CowUnavailable(message) => Error::CowUnavailable(format!(
+            "{} does not support Linux copy-on-write reflinks: {message}",
+            path.display()
+        )),
+        error => error,
+    });
+    let cleanup = [&source, &destination]
+        .into_iter()
+        .filter(|path| path.exists())
+        .try_for_each(fs::remove_file);
+    result.and(cleanup.map_err(Error::from))
 }
 
 #[derive(Clone, Copy)]
@@ -197,4 +252,137 @@ pub(super) fn c_path(path: &Path) -> Result<std::ffi::CString> {
 
     std::ffi::CString::new(path.as_os_str().as_bytes())
         .map_err(|_| Error::Path(format!("path contains a null byte: {}", path.display())))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::strategy::linux::{Filesystem, LinuxStrategy, filesystem};
+    use std::os::unix::fs::{MetadataExt, PermissionsExt};
+    use tempfile::{Builder, TempDir};
+
+    fn reflink_temp() -> Option<TempDir> {
+        let temp = Builder::new()
+            .prefix(".rift-core-test-")
+            .tempdir_in(std::env::current_dir().unwrap())
+            .unwrap();
+        (!matches!(filesystem(temp.path()).unwrap(), Filesystem::Btrfs)
+            && verify_reflinks_linux(temp.path()).is_ok())
+        .then_some(temp)
+    }
+
+    fn non_reflink_temp() -> Option<TempDir> {
+        let temp = TempDir::new().unwrap();
+        (!matches!(filesystem(temp.path()).unwrap(), Filesystem::Btrfs)
+            && verify_reflinks_linux(temp.path()).is_err())
+        .then_some(temp)
+    }
+
+    #[test]
+    fn linux_reflink_integration_environment_is_available() {
+        if std::env::var_os("RIFT_REQUIRE_REFLINK_TESTS").is_some() {
+            assert!(
+                reflink_temp().is_some(),
+                "RIFT_REQUIRE_REFLINK_TESTS requires a non-btrfs Linux filesystem with reflinks"
+            );
+        }
+    }
+
+    #[test]
+    fn native_init_verifies_reflink_support() {
+        let Some(temp) = reflink_temp() else {
+            return;
+        };
+        assert_eq!(
+            LinuxStrategy
+                .initialize_directory(temp.path(), &mut |_| {})
+                .unwrap(),
+            StrategyInit::AlreadyNative
+        );
+    }
+
+    #[test]
+    fn native_copy_preserves_files_links_and_metadata() {
+        let Some(temp) = reflink_temp() else {
+            return;
+        };
+        let source = temp.path().join("source");
+        let destination = temp.path().join("destination");
+        let nested = source.join("nested");
+        fs::create_dir(&source).unwrap();
+        fs::set_permissions(&source, fs::Permissions::from_mode(0o750)).unwrap();
+        fs::create_dir(&nested).unwrap();
+        let file = nested.join("file.txt");
+        fs::write(&file, "hello").unwrap();
+        fs::set_permissions(&file, fs::Permissions::from_mode(0o640)).unwrap();
+        fs::hard_link(&file, nested.join("hard.txt")).unwrap();
+        std::os::unix::fs::symlink("file.txt", nested.join("link.txt")).unwrap();
+
+        LinuxStrategy.copy_directory(&source, &destination).unwrap();
+
+        assert_eq!(
+            fs::read_to_string(destination.join("nested/file.txt")).unwrap(),
+            "hello"
+        );
+        assert_eq!(
+            fs::read_link(destination.join("nested/link.txt")).unwrap(),
+            Path::new("file.txt")
+        );
+        assert_eq!(
+            fs::metadata(destination.join("nested/file.txt"))
+                .unwrap()
+                .ino(),
+            fs::metadata(destination.join("nested/hard.txt"))
+                .unwrap()
+                .ino()
+        );
+        assert_eq!(
+            fs::metadata(destination.join("nested/file.txt"))
+                .unwrap()
+                .permissions()
+                .mode()
+                & 0o777,
+            0o640
+        );
+        assert_eq!(
+            fs::metadata(&destination).unwrap().permissions().mode() & 0o777,
+            0o750
+        );
+        LinuxStrategy.remove_directory(&destination).unwrap();
+        assert!(!destination.exists());
+    }
+
+    #[test]
+    fn native_copy_rejects_storage_on_another_filesystem() {
+        let Some(temp) = reflink_temp() else {
+            return;
+        };
+        let source = temp.path().join("source");
+        fs::create_dir(&source).unwrap();
+        let other = TempDir::new().unwrap();
+        if fs::metadata(&source).unwrap().dev() == fs::metadata(other.path()).unwrap().dev() {
+            return;
+        }
+
+        assert!(matches!(
+            LinuxStrategy.copy_directory(&source, &other.path().join("destination")),
+            Err(Error::CowUnavailable(_))
+        ));
+    }
+
+    #[test]
+    fn native_copy_rejects_empty_trees_without_reflink_support() {
+        let Some(temp) = non_reflink_temp() else {
+            return;
+        };
+        let source = temp.path().join("source");
+        let destination = temp.path().join("destination");
+        fs::create_dir(&source).unwrap();
+
+        assert!(matches!(
+            LinuxStrategy.copy_directory(&source, &destination),
+            Err(Error::CowUnavailable(_))
+        ));
+        assert!(!destination.exists());
+    }
 }
